add testPrint.cpp
