+++ config_file/default.conf
@@ -1,6 +1,7 @@
 server {
   listen 80;
   host localhost;
+  server_names www.webserv.com webserv.com;
     error_page 500 500.html;
     error_page 403 403.html;
     client_max_body_size 1000000;
