+++ src/CMakeLists.txt
@@ -15,6 +15,7 @@ add_library(webserv_lib STATIC
     config/config.cpp
     config/parser.cpp
     config/tokenizer.cpp
+    config/testPrint.cpp
     config/context/server.cpp
 	config/context/location.cpp
     config/context/document_root.cpp
